+++ sdk-typescript/.lintstagedrc.json
@@ -1,0 +1,4 @@
+{
+  "*": "prettier --ignore-unknown --write",
+  "*.ts": "eslint --fix"
+}
