--- sdk-typescript/models/IndexDistance.ts
@@ -1,9 +1,0 @@
-/* istanbul ignore file */
-/* tslint:disable */
-/* eslint-disable */
-
-export enum IndexDistance {
-    DOT = 'dot',
-    COSINE = 'cosine',
-    EUCLIDEAN = 'euclidean',
-}
