--- sdk-typescript/models/SearchRequest.ts
@@ -1,10 +1,0 @@
-/* istanbul ignore file */
-/* tslint:disable */
-/* eslint-disable */
-
-export type SearchRequest = {
-    index: string;
-    'k'?: number | null;
-    query: string;
-};
-
