--- sdk-typescript/models/TextSplitterKind.ts
@@ -1,13 +1,0 @@
-/* istanbul ignore file */
-/* tslint:disable */
-/* eslint-disable */
-
-export type TextSplitterKind = ('none' | 'new_line' | {
-    /**
-     * Split a document across the regex boundary
-     */
-    regex: {
-        pattern: string;
-    };
-});
-
