+++ sdk-typescript/src/models/DataConnector.ts
@@ -1,10 +1,9 @@
 /* istanbul ignore file */
 /* tslint:disable */
-/* eslint-disable */
+ 
 
 import type { SourceType } from './SourceType';
 
 export type DataConnector = {
-    source: SourceType;
+  source: SourceType;
 };
-
