+++ sdk-typescript/src/models/DocumentFragment.ts
@@ -3,7 +3,6 @@
 /* eslint-disable */
 
 export type DocumentFragment = {
-    metadata: Record<string, any>;
-    text: string;
+  metadata: Record<string, any>;
+  text: string;
 };
-
