+++ sdk-typescript/src/models/Extractor.ts
@@ -1,13 +1,12 @@
 /* istanbul ignore file */
 /* tslint:disable */
-/* eslint-disable */
+ 
 
 import type { ExtractorContentType } from './ExtractorContentType';
 import type { ExtractorType } from './ExtractorType';
 
 export type Extractor = {
-    content_type: ExtractorContentType;
-    extractor_type: ExtractorType;
-    name: string;
+  content_type: ExtractorContentType;
+  extractor_type: ExtractorType;
+  name: string;
 };
-
