+++ sdk-typescript/src/models/ExtractorContentType.ts
@@ -1,8 +1,8 @@
 /* istanbul ignore file */
 /* tslint:disable */
-/* eslint-disable */
+ 
 
 export enum ExtractorContentType {
-    TEXT = 'text',
-    MEMORY = 'memory',
+  TEXT = 'text',
+  MEMORY = 'memory',
 }
