+++ sdk-typescript/src/models/ExtractorType.ts
@@ -1,15 +1,14 @@
 /* istanbul ignore file */
 /* tslint:disable */
-/* eslint-disable */
+ 
 
 import type { IndexDistance } from './IndexDistance';
 import type { TextSplitterKind } from './TextSplitterKind';
 
 export type ExtractorType = {
-    embedding: {
-        distance: IndexDistance;
-        model: string;
-        text_splitter: TextSplitterKind;
-    };
+  embedding: {
+    distance: IndexDistance;
+    model: string;
+    text_splitter: TextSplitterKind;
+  };
 };
-
