+++ sdk-typescript/src/models/IndexDistance.ts
@@ -1,0 +1,9 @@
+/* istanbul ignore file */
+/* tslint:disable */
+ 
+
+export enum IndexDistance {
+  DOT = 'dot',
+  COSINE = 'cosine',
+  EUCLIDEAN = 'euclidean',
+}
