+++ sdk-typescript/src/models/IndexSearchResponse.ts
@@ -1,10 +1,9 @@
 /* istanbul ignore file */
 /* tslint:disable */
-/* eslint-disable */
+ 
 
 import type { DocumentFragment } from './DocumentFragment';
 
 export type IndexSearchResponse = {
-    results: Array<DocumentFragment>;
+  results: Array<DocumentFragment>;
 };
-
