+++ sdk-typescript/src/models/SearchRequest.ts
@@ -1,0 +1,9 @@
+/* istanbul ignore file */
+/* tslint:disable */
+ 
+
+export type SearchRequest = {
+  index: string;
+  k?: number | null;
+  query: string;
+};
