+++ sdk-typescript/src/models/SourceType.ts
@@ -1,14 +1,15 @@
 /* istanbul ignore file */
 /* tslint:disable */
-/* eslint-disable */
+ 
 
-export type SourceType = ({
-    google_contact: {
+export type SourceType =
+  | {
+      google_contact: {
         metadata?: string | null;
-    };
-} | {
-    gmail: {
+      };
+    }
+  | {
+      gmail: {
         metadata?: string | null;
+      };
     };
-});
-
