+++ sdk-typescript/src/models/SyncRepository.ts
@@ -6,9 +6,8 @@ import type { DataConnector } from './DataConnector';
 import type { Extractor } from './Extractor';
 
 export type SyncRepository = {
-    data_connectors: Array<DataConnector>;
-    extractors: Array<Extractor>;
-    metadata: Record<string, any>;
-    name: string;
+  data_connectors: Array<DataConnector>;
+  extractors: Array<Extractor>;
+  metadata: Record<string, any>;
+  name: string;
 };
-
