+++ sdk-typescript/src/models/Text.ts
@@ -3,7 +3,6 @@
 /* eslint-disable */
 
 export type Text = {
-    metadata: Record<string, any>;
-    text: string;
+  metadata: Record<string, any>;
+  text: string;
 };
-
