+++ sdk-typescript/src/models/TextAddRequest.ts
@@ -1,11 +1,10 @@
 /* istanbul ignore file */
 /* tslint:disable */
-/* eslint-disable */
+ 
 
 import type { Text } from './Text';
 
 export type TextAddRequest = {
-    documents: Array<Text>;
-    repository?: string | null;
+  documents: Array<Text>;
+  repository?: string | null;
 };
-
