+++ sdk-typescript/src/models/TextSplitterKind.ts
@@ -1,0 +1,15 @@
+/* istanbul ignore file */
+/* tslint:disable */
+ 
+
+export type TextSplitterKind =
+  | 'none'
+  | 'new_line'
+  | {
+      /**
+       * Split a document across the regex boundary
+       */
+      regex: {
+        pattern: string;
+      };
+    };
