+++ sdk-typescript/tsconfig.module.json
@@ -1,0 +1,9 @@
+{
+  "extends": "./tsconfig",
+  "compilerOptions": {
+    "target": "esnext",
+    "outDir": "lib/module",
+    "module": "esnext"
+  },
+  "exclude": ["node_modules/**"]
+}
